backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.28.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
qdrant-client>=1.9.0
rank-bm25>=0.2.2
tqdm>=4.66.0
pypdf>=4.0.0
python-multipart>=0.0.9
